Refactor explore section into its own file.
